+++ WebAppOppg2/Controllers/RouteController.cs
@@ -12,7 +12,7 @@ namespace WebAppOppg2.RouteController
 {
     [ApiController]
 
-    [Route ("api/[controller]")] 
+    [Route("api/[controller]")]
 
     public class RouteController : ControllerBase
     {
@@ -27,46 +27,41 @@ namespace WebAppOppg2.RouteController
             _log = log;
         }
 
-        //SaveTicket
         [HttpPost]
         public async Task<ActionResult> SaveRoute(Route route)
         {
-            /*
             if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
             {
-                return Unauthorized();
-            }*/
+                return Unauthorized("Rute ikke lagret - status: ikke innlogget");
+            }
             if (ModelState.IsValid)
             {
                 bool returOK = await _db.AddRoute(route);
                 if (!returOK)
                 {
                     _log.LogInformation("Rute kunne ikke lagres!");
-                    return BadRequest("");
+                    return BadRequest("Rute ikke lagret, feil i DB - status: innlogget");
                 }
-                return Ok("");
+                return Ok("Rute lagret - status: innlogget");
             }
             _log.LogInformation("Feil i inputvalidering");
-            return BadRequest("");
+            return BadRequest("Rute ikke lagret, feil ved inputvalideringen på server - status: innlogget");
         }
 
-
-        //GetOne
         [HttpGet("{id}")]
         public async Task<ActionResult> GetOneRoute(int id)
         {
-            /*
             if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
             {
-                return Unauthorized();
-            }*/
+                return Unauthorized("Rute ikke hentet, feil i DB - status: ikke innlogget");
+            }
             if (ModelState.IsValid)
             {
                 Route route = await _db.GetOneRoute(id);
                 if (route == null)
                 {
                     _log.LogInformation("Fant ikke rute");
-                    return NotFound();
+                    return NotFound("Rute ikke hentet, feil i DB - status: innlogget");
                 }
                 return Ok(route);
             }
@@ -74,55 +69,52 @@ namespace WebAppOppg2.RouteController
             return BadRequest();
         }
 
-        //GetALl
         [HttpGet]
         public async Task<ActionResult> GetAllRoutes()
         {
-            /*
             if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
             {
-                return Unauthorized();
+                return Unauthorized("Ruter ikke hentet, feil i DB - status: ikke innlogget");
             }
-            */
             List<Route> routes = await _db.GetAllRoutes();
             return Ok(routes);
         }
 
-        //DeleteTicket
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteRoute(int id)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
+            {
+                return Unauthorized("Rute ikke slettet, feil i DB - status: ikke innlogget");
+            }
             bool returOK = await _db.DeleteRoute(id);
             if (!returOK)
             {
                 _log.LogInformation("Sletting av ruten ble ikke utført");
-                return NotFound();
+                return NotFound("Rute ikke slettet, feil i DB - status: innlogget");
             }
-            return Ok();
-
+            return Ok("Rute slettet - status: innlogget");
         }
 
-        //EditTicket
         [HttpPut]
         public async Task<ActionResult> EditRoute(Route route)
         {
-            /*
             if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
             {
-                return Unauthorized();
-            }*/
+                return Unauthorized("Rute ikke endret - status: ikke innlogget");
+            }
             if (ModelState.IsValid)
             {
                 bool returOK = await _db.EditRoute(route);
                 if (!returOK)
                 {
                     _log.LogInformation("Endringen kunne ikke utføres");
-                    return NotFound();
+                    return NotFound("Rute ikke endret, feil i DB - status: innlogget");
                 }
-                return Ok();
+                return Ok("Rute endret - status: innlogget");
             }
             _log.LogInformation("Feil i inputvalidering");
-            return BadRequest();
+            return BadRequest("Rute ikke endret, feil ved inputvalideringen på server - status: innlogget");
         }
     }
 }
