+++ WebAppOppg2/Controllers/TravelTypeController.cs
@@ -12,7 +12,7 @@ namespace WebAppOppg2.TravelTypeController
 {
     [ApiController]
 
-    [Route ("api/[controller]")]
+    [Route("api/[controller]")]
 
     public class TravelTypeController : ControllerBase
     {
@@ -27,46 +27,41 @@ namespace WebAppOppg2.TravelTypeController
             _log = log;
         }
 
-        //SaveTicket
         [HttpPost]
         public async Task<ActionResult> SaveTravelType(TravelType travelType)
         {
-            /*
             if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
             {
-                return Unauthorized();
-            }*/
+                return Unauthorized("Reisetype ikke lagret - status: ikke innlogget");
+            }
             if (ModelState.IsValid)
             {
                 bool returOK = await _db.AddTravelType(travelType);
                 if (!returOK)
                 {
                     _log.LogInformation("Traveltype kunne ikke lagres!");
-                    return BadRequest("");
+                    return BadRequest("Reisetype ikke lagret, feil i DB - status: innlogget");
                 }
-                return Ok("");
+                return Ok("Reisetype lagret - status: innlogget");
             }
             _log.LogInformation("Feil i inputvalidering");
-            return BadRequest("");
+            return BadRequest("Reisetype ikke lagret, feil ved inputvalideringen på server - status: innlogget");
         }
 
-
-        //GetOne
         [HttpGet("{id}")]
         public async Task<ActionResult> GetOneTravelType(int id)
         {
-            /*
             if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
             {
-                return Unauthorized();
-            }*/
+                return Unauthorized("Reisetype ikke hentet, feil i DB - status: ikke innlogget");
+            }
             if (ModelState.IsValid)
             {
                 TravelType travelType = await _db.GetOneTravelType(id);
                 if (travelType == null)
                 {
                     _log.LogInformation("Fant ikke traveltype");
-                    return NotFound();
+                    return NotFound("Reisetype ikke hentet, feil i DB - status: innlogget");
                 }
                 return Ok(travelType);
             }
@@ -74,55 +69,53 @@ namespace WebAppOppg2.TravelTypeController
             return BadRequest();
         }
 
-        //GetALl
         [HttpGet]
         public async Task<ActionResult> GetAllTravelTypes()
         {
-            /*
             if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
             {
-                return Unauthorized();
+                return Unauthorized("Reisetype ikke hentet, feil i DB - status: ikke innlogget");
             }
-            */
             List<TravelType> travelTypes = await _db.GetAllTravelTypes();
             return Ok(travelTypes);
         }
 
-        //DeleteTicket
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteTravelType(int id)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
+            {
+                return Unauthorized("Reisetype ikke slettet, feil i DB - status: ikke innlogget");
+            }
             bool returOK = await _db.DeleteTravelType(id);
             if (!returOK)
             {
                 _log.LogInformation("Sletting av traveltype ble ikke utført");
-                return NotFound();
+                return NotFound("Reisetype ikke slettet, feil i DB - status: innlogget");
             }
-            return Ok();
+            return Ok("Reisetype slettet - status: innlogget");
 
         }
 
-        //EditTicket
         [HttpPut]
         public async Task<ActionResult> EditTravelType(TravelType travelType)
         {
-            /*
             if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggetInn)))
             {
-                return Unauthorized();
-            }*/
+                return Unauthorized("Reisetype ikke endret - status: ikke innlogget");
+            }
             if (ModelState.IsValid)
             {
                 bool returOK = await _db.EditTravelType(travelType);
                 if (!returOK)
                 {
                     _log.LogInformation("Endringen kunne ikke utføres");
-                    return NotFound();
+                    return NotFound("Reisetype ikke endret, feil i DB - status: innlogget");
                 }
-                return Ok();
+                return Ok("Reisetype endret - status: innlogget");
             }
             _log.LogInformation("Feil i inputvalidering");
-            return BadRequest();
+            return BadRequest("Reisetype ikke endret, feil ved inputvalideringen på server - status: innlogget");
         }
     }
 }
